requirements.txt
pytest>=8.0.0
playwright>=1.49.0
pytest-playwright>=0.6.2
allure-pytest>=2.13.5
